telenor_health/requirements.txt

Django==1.11
djangorestframework
django-redshift-backend